Support Windows users with .sh postgre startup script \r auto-removal
Using docker-compose volumes to map .sh scripts into container leads to a "trailing \r problem" (script can't be executed) when running the container from a windows machine. Which leads to curropted environment and problems for Windows users to run this demo.

Instead of using docker-compose volumes, I have created a new Docker files that extends the postgre image in order to remove the trailing \r when adding this file into the container using dos2unix utility.