Use .SQL file instead of .sh file to initialize db data.
Changing the way of loading data into the DATABASE from .sh file to .sql file - The main reason to do that is because .sh files that exposed through docker-compose into the container can cause "trailing \r character" problem for windows users when they are being executed by the container. Therefore, instead of using a .sh file to initiate the data into the DB, we use a simple .SQL file that will be exposed to the db conainer, and will be executed by MySQL when loaded.